third_party/blink/web_tests/fast/html/meter-user-modify.html

<!DOCTYPE html>
<html>
<head>
<style>
meter::-webkit-meter-bar {
    -webkit-user-modify: read-write;
}

meter {
    -webkit-appearance: none;
}
</style>
</head>
<body>
<script src="../../resources/js-test.js"></script>

<p>The meter content doesn't get rendered.</p>

<div id="container" contenteditable>
    (before)<meter id="meter" min="0" max="100" value="50">In meter</meter>(after)
</div>

<script>
function focusAndType(id, key)
{
    var target = document.getElementById(id);
    eventSender.mouseMoveTo(target.offsetLeft + target.offsetWidth / 2, target.offsetTop + target.offsetHeight / 2);
    eventSender.mouseDown();
    eventSender.mouseUp();
    eventSender.keyDown(key);
}

var container = document.getElementById("container");
layoutTreeBeforeType = internals.elementLayoutTreeAsText(document.documentElement);

focusAndType("meter", "X");

layoutTreeAfterType = internals.elementLayoutTreeAsText(document.documentElement);
shouldBe("layoutTreeAfterType", "layoutTreeBeforeType");

var successfullyParsed = true;
</script>
</body>
</html>
